% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pa.R
\name{calculate_CEAC}
\alias{calculate_CEAC}
\title{Function to calculate the cost-effectiveness acceptability curves.}
\usage{
calculate_CEAC(Q.trt, C.trt, Q.comp, C.comp, v.wtp)
}
\arguments{
\item{Q.trt}{a vector of numeric. Contains the mean QALYs per individual for the intervention strategy.}

\item{C.trt}{a vector of numeric. Contains the mean costs per individual for the intervention strategy.}

\item{Q.comp}{a vector of numeric. Contains the mean QALYs per individual for the comparator strategy.}

\item{C.comp}{a vector of numeric. Contains the mean costs per individual for the comparator strategy.}

\item{v.wtp}{a vector. This vector contains the different willingness to pay thresholds.}
}
\value{
A matrix.
}
\description{
This function calculates the probability that each strategy is the most cost effective at different willingness-to-pay threshold.
}
\details{
This function has been obtained from: XXX.
}
