% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pa.R
\name{generate_pa_inputs}
\alias{generate_pa_inputs}
\title{Function to generate probabilistic inputs}
\usage{
generate_pa_inputs(
  n_sim = n_sim,
  mean_eff_mi = 0.6,
  se_eff_mi = 0.15,
  seed_num = 452
)
}
\arguments{
\item{n_sim}{integer. Determine the number of probabilistic input parameters to define.}

\item{mean_eff_mi}{numeric. Determine the mean relative effectiveness of the intervention on the occurence of MI.}

\item{se_eff_mi}{numeric. Determine the standard error surrounding the mean relative effectiveness of the intervention on the occurence of MI.}

\item{seed_num}{integer. Seed number to use to generate the probabilistic input parameters.}
}
\value{
A dataframe.
}
\description{
This function generate the probabilistic input parameters for the health economic model
}
