% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_pa.R
\name{perform_simulation}
\alias{perform_simulation}
\title{Function to perform the health economic analysis.}
\usage{
perform_simulation(l_params, start_age = 45, verbose = FALSE)
}
\arguments{
\item{l_params}{list. List of input parameters.}

\item{start_age}{integer. Determine the mean age of the cohort at the start of the health economic model.}

\item{verbose}{logical. Default is FALSE.}
}
\value{
A vector.
}
\description{
This function performs the analysis using the example health economic model.
}
